requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
plotly>=5.8.0
pillow>=9.0.0
